FEAT: CRUD methods added to crudable. Pass tests in M-R
